+++ src/frontend/src/instance/log_handler.rs
@@ -84,6 +84,10 @@ impl PipelineHandler for Instance {
         Ok(outputs)
     }
 
+    fn check_pipeline_query_permission(&self, query_ctx: &QueryContextRef) -> ServerResult<()> {
+        self.check_permission(query_ctx, PermissionReq::PipelineQuery)
+    }
+
     async fn get_pipeline(
         &self,
         name: &str,
@@ -103,6 +107,7 @@ impl PipelineHandler for Instance {
         pipeline: &str,
         query_ctx: QueryContextRef,
     ) -> ServerResult<PipelineInfo> {
+        self.check_permission(&query_ctx, PermissionReq::PipelineManage)?;
         self.pipeline_operator
             .insert_pipeline(name, content_type, pipeline, query_ctx)
             .await
@@ -115,6 +120,7 @@ impl PipelineHandler for Instance {
         version: PipelineVersion,
         ctx: QueryContextRef,
     ) -> ServerResult<Option<()>> {
+        self.check_permission(&ctx, PermissionReq::PipelineManage)?;
         self.pipeline_operator
             .delete_pipeline(name, version, ctx)
             .await
@@ -143,6 +149,7 @@ impl PipelineHandler for Instance {
         version: PipelineVersion,
         query_ctx: QueryContextRef,
     ) -> ServerResult<(String, TimestampNanosecond)> {
+        self.check_permission(&query_ctx, PermissionReq::PipelineQuery)?;
         self.pipeline_operator
             .get_pipeline_str(name, version, query_ctx)
             .await
